requirements.txt
numpy
opencv-python
opencv-python-headless
kraken
Pillow
qreader
pandas
scipy
imutils
dbr